requirements.txt
aiohttp==3.9.0
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
charset-normalizer==3.3.0
envparse==0.2.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
six==1.16.0
typing_extensions==4.8.0
webencodings==0.5.1
yarl==1.9.2
